uncompressed: separated codec from ImageItem
